Order,Book Controller

+++ src/main/java/com/bookstore/project/repository/BookRepository.java
@@ -1,15 +1,12 @@
 package com.bookstore.project.repository;
 
-import com.bookstore.project.dto.BookDTO;
 import com.bookstore.project.entity.Book;
-import com.bookstore.project.entity.Genre;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BookRepository extends JpaRepository<Book, Integer> {
-    List<Book> findByGenreId(int genreId);
+public interface BookRepository extends JpaRepository<Book, Long> {
+    List<Book> findByGenreId(long genreId);
     List<Book> findByUserId(long userId);
 }
